InfoMenu: - Image is now visible - Animal Color is now default and will need a new Texture

+++ Assets/Scripts/Util.cs
@@ -22,8 +22,8 @@ public static class Util
     }
     public static Vector2 getRandomCoordinateInPlayground()
     {
-        int halfW = (int)Gamevariables.playgroundSize.x/2;
-        int halfH = (int)Gamevariables.playgroundSize.y/2;
+        int halfW = Gamevariables.playgroundSize.x/2;
+        int halfH = Gamevariables.playgroundSize.y/2;
         return new Vector2(Random.Range(-halfW, halfW), Random.Range(-halfH, halfH));
     }
